fix(task runner): try double fork hack

+++ lib/task_runner.py
@@ -35,24 +35,28 @@ class TaskRunner():
             raise FileExistsError('%s task is in progress' % self.taskname)
         child_pid = os.fork()
         if child_pid == 0:
-            # child process
             os.setsid()
-            task_run_id = self.get_human_readable_timestamp()
+            inner_child_pid = os.fork()
+            if inner_child_pid == 0:
+                # child process
+                task_run_id = self.get_human_readable_timestamp()
 
-            self.history_storage.prepare_task_run(self.taskname, task_run_id)
+                self.history_storage.prepare_task_run(self.taskname, task_run_id)
 
-            self.state['current'] = 'running'
-            self.state['started_at'] = self.get_timestamp()
-            self.history_storage.update_state_for_task_run_id(self.taskname, task_run_id, self.state)
-            self.history_storage.update_current_task_run_for_task(self.taskname, task_run_id)
-            proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-            self.history_storage.update_pid_for_task_run_id(self.taskname, task_run_id, str(proc.pid))
-            stdout, stderr = proc.communicate()
-            self.state['exit_code'] = proc.returncode
-            self.history_storage.update_stdout_for_task_run_id(self.taskname, task_run_id, stdout.decode())
-            self.history_storage.update_stderr_for_task_run_id(self.taskname, task_run_id, stderr.decode())
-            self.state['finished_at'] = self.get_timestamp()
-            self.state['current'] = 'finished'
-            self.history_storage.update_state_for_task_run_id(self.taskname, task_run_id, self.state)
-            self.history_storage.delete_current_task_run_for_task(self.taskname)
-            sys.exit(0)
+                self.state['current'] = 'running'
+                self.state['started_at'] = self.get_timestamp()
+                self.history_storage.update_state_for_task_run_id(self.taskname, task_run_id, self.state)
+                self.history_storage.update_current_task_run_for_task(self.taskname, task_run_id)
+                proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+                self.history_storage.update_pid_for_task_run_id(self.taskname, task_run_id, str(proc.pid))
+                stdout, stderr = proc.communicate()
+                self.state['exit_code'] = proc.returncode
+                self.history_storage.update_stdout_for_task_run_id(self.taskname, task_run_id, stdout.decode())
+                self.history_storage.update_stderr_for_task_run_id(self.taskname, task_run_id, stderr.decode())
+                self.state['finished_at'] = self.get_timestamp()
+                self.state['current'] = 'finished'
+                self.history_storage.update_state_for_task_run_id(self.taskname, task_run_id, self.state)
+                self.history_storage.delete_current_task_run_for_task(self.taskname)
+                sys.exit(0)
+            else:
+                sys.exit(0)
